[fix] チャットが送れないバグの修正
